did a shit load of stuff -AI management start stuff -Flow fields now work -Darkness thing with flow fields -Started behaviour tree -Started sound (game logic, not actual audio) implementation

+++ project2D/PhysObject.cpp
@@ -47,7 +47,7 @@ void PhysObject::Update()
 		velocity = velocity.Normalised() * MAX_SPEED;
 
 	//Update position
-	entity->GetTransform()->Rotate(angularVelocity * DEG2RAD * deltaTime);
+	entity->GetTransform()->Rotate(angularVelocity * (float)DEG2RAD * deltaTime);
 	entity->GetTransform()->Translate(velocity * deltaTime);
 }
